[#43697] fixed PR issues
- added storage icon factory
- renaming of view model classes and variables
- fixes of spot variable usage
