+++ frontend/src/app/shared/components/file-links/file-link-list/file-link-list.component.ts
@@ -155,7 +155,8 @@ export class FileLinkListComponent extends UntilDestroyedMixin implements OnInit
 
   public openLinkFilesDialog():void {
     const locals = {
-      storageType: this.storageType,
+      storageType: this.storage._links.type.href,
+      storageName: this.storage.name,
       storageLocation: this.storageFilesLocation,
       storageLink: this.storage._links.self,
     };
