+++ frontend/src/app/shared/components/file-links/file-picker-modal/file-picker-modal.html
@@ -17,7 +17,7 @@
 
   <div class="spot-modal--body spot-container">
     <ng-container *ngIf="(loading$ | async) === false; else loadingTemplate">
-      <op-breadcrumbs [viewModel]="breadCrumbs"></op-breadcrumbs>
+      <op-breadcrumbs [content]="breadcrumbs"></op-breadcrumbs>
 
       <ul
         class="spot-list spot-list_compact op-file-list"
@@ -47,16 +47,17 @@
         [textContent]="text.buttons.cancel"
         (click)="closeMe()"
       ></button>
-      <button
-        type="button"
-        class="button spot-action-bar--action"
-        [ngClass]="{ '-highlight': selectedFileCount > 0 }"
-        [disabled]="selectedFileCount === 0"
-        (click)="createSelectedFileLinks()"
-      >
-        <span class="spot-icon spot-icon_add-link"></span>
-        <span>{{selectedFileCount > 0 ? text.buttons.submit() : text.buttons.submitEmptySelection}}</span>
-      </button>
+<!--      TODO: #44376, enable button with functionality-->
+<!--      <button-->
+<!--        type="button"-->
+<!--        class="button spot-action-bar&#45;&#45;action"-->
+<!--        [ngClass]="{ '-highlight': selectedFileCount > 0 }"-->
+<!--        [disabled]="selectedFileCount === 0"-->
+<!--        (click)="createSelectedFileLinks()"-->
+<!--      >-->
+<!--        <span class="spot-icon spot-icon_add-link"></span>-->
+<!--        <span>{{selectedFileCount > 0 ? text.buttons.submit() : text.buttons.submitEmptySelection}}</span>-->
+<!--      </button>-->
     </div>
   </div>
 </div>
