+++ frontend/src/app/spot/styles/sass/components/breadcrumbs.sass
@@ -11,8 +11,8 @@
     min-width: 0
 
     > .spot-icon
-      width: $spot-spacing-1-5
-      height: $spot-spacing-1-5
+      width: $spot-spacing-1_5
+      height: $spot-spacing-1_5
 
     &_shrink
       flex-shrink: 1
@@ -24,7 +24,7 @@
       text-overflow: ellipsis
 
       &:not(:first-child)
-        margin-left: $spot-spacing-0-25
+        margin-left: $spot-spacing-0_25
 
 
     &-divider
@@ -32,4 +32,4 @@
       flex-shrink: 0
 
     &:not(:first-child)
-      margin-left: $spot-spacing-0-25
+      margin-left: $spot-spacing-0_25
